single list updated

+++ shoppinglistanalyzer/src/data_structures.rs
@@ -1,13 +1,13 @@
 #[derive(Clone)]
 pub struct Item {
-    pub id: i32,
+    pub id: i64,
     pub name: String,
     pub category: String,
-    pub price: f32,
+    pub price: f64,
 }
 
 impl Item {
-    pub fn new(identifier: i32, item_name: String, cat: String, m_price: f32) -> Self {
+    pub fn new(identifier: i64, item_name: String, cat: String, m_price: f64) -> Self {
         Self {
             id: identifier,
             name: item_name,
@@ -42,8 +42,8 @@ impl List {
         }
     }
 
-    pub fn get_total_cost(&self) -> f32 {
-        let mut total: f32 = 0.0;
+    pub fn get_total_cost(&self) -> f64 {
+        let mut total: f64 = 0.0;
         for item in self.items.iter() {
             total += item.price;
         }
